feat: add project metadata
