+++ internal/providers/terraform/dir_provider.go
@@ -96,10 +96,6 @@ func (p *DirProvider) checks() error {
 }
 
 func (p *DirProvider) AddMetadata(metadata *schema.ProjectMetadata) {
-	if metadata.TerraformWorkspace != "" {
-		return
-	}
-
 	terraformWorkspace := p.Workspace
 
 	if terraformWorkspace == "" {
